Serial comms problem with Arduino Uno Rev2/3
I had to alter the serial comms check and add a delay after creating the
serial connection otherwise it would sit in a loop forever.  The output
is also a little more verbose just so I could see what was going on.

I added the delay after creating the serial connection as the act of
creating a serial connection resets the Arduino Uno (tested Rev2 and
Rev3 boards). Then the gcode is sent before the Arduino is ready to
receive it and hence doesn't respond with 'ok:' and we never break out
of the loop even when readline() times out. I altered the way the
readline() checks are done so if a time out occurs the execution
continues and sends the next line from the gcode file.

I found the changes to the readline() data checking aren't necessary if
the serial connection auto restart delay is sufficently long then the
original code works as expected. However I wasn't sure if the complexity
of the Arduino firmware affects its startup time, so I am leaving it in
as it allows the gcode to do something even if the serial connection
auto restart delay isn't long enough. This is based on the assumption
that the comms issue is at start up as the first line of a gcode file
created by the Unicorn gcode plugin for Inkscape is a comment and so
nothing is lost if the first line is never received by the Arduino.